\name{eq}
\alias{eq}
\docType{data}
\title{
Earthquake dataset as \code{"stpp"} object
}
\description{
Locations and times of earthquakes of magnitude 3.5 and up, from 1/1/2002 - 12/10/2010.
}
\usage{data(eq)}
\format{
A \code{"stpp"} object
}
\details{
Times are measured in minutes from first earthquake in the dataset. 
}
\source{
\url{http://www.ncedc.org/anss/catalog-search.html}
}

\examples{
data(eq)
}
\keyword{datasets}
