\name{print.stgrid}
\alias{print.stgrid}
\title{
Print details of a \code{stgrid} object
}
\description{
Prints a full spatial grid.
}
\usage{
print.stgrid(x)
}
\arguments{
  \item{x}{
A \code{\dQuote{stgrid}} object.
}
}
\author{
Robert Clements
}
\seealso{
\code{\link{make.grid}}
}
\examples{
data(gresiduals)
gresiduals$grid
}
