\name{print.stpp}
\alias{print.stpp}
\title{
Print details of a \code{stpp} object
}
\description{
Prints the x, y, and t coordinates of a space-time dataset, and prints the limits of the space-time window.
}
\usage{
print.stpp(X)
}
\arguments{
  \item{X}{
A \code{\dQuote{stpp}} object.
}
}
\author{
Robert Clements
}
\seealso{
\code{\link{stpp}}
}
\examples{
data(eq)
eq
}
