\name{redbanana}
\alias{redbanana}
\docType{data}
\title{
Space-time data for red banana plants
}
\description{
Spatial locations and birth times of 788 red banana plants. Locations given in longitude and latitude.
}
\usage{data(redbanana)}
\format{
  A data frame with 788 observations on the following 3 variables.
  \describe{
    \item{\code{longitude}}{a numeric vector}
    \item{\code{latitude}}{a numeric vector}
    \item{\code{birth}}{a numeric vector}
  }
}
\details{
}
\source{
Balderama, E., Schoenberg, F.P., Murray, E., and Rundel, P.W. (2012) Application of branching point process models to the study of invasive red banana plants in Costa Rica. \emph{Journal of the American Statistical Association,} to appear. 
}
\references{
}
\examples{
data(redbanana)
}
\keyword{datasets}
