\name{print.gridresid}
\alias{print.gridresid}
\title{
Print details of a \code{gridresid} object
}
\description{
Prints the details of the results of finding grid-based residuals for a space-time point process.
}
\usage{
print.gridresid(x)
}
\arguments{
  \item{x}{
A \code{\dQuote{gridresid}} object.
}
}
\details{
Prints the details of the space-time dataset, the grid-based residuals, and the spatial grid.
}
\author{
Robert Clements
}
\seealso{
\code{\link{gridresid}}
}
\examples{
data(gresiduals)
gresiduals
}
