\name{print.stpp}
\alias{print.stpp}
\title{
Print details of a \code{stpp} object
}
\description{
Prints the x, y, and t coordinates of a space-time dataset, and prints the limits of the space-time window.
}
\usage{
\method{print}{stpp}(x, \ldots)
}
\arguments{
  \item{x}{
A \dQuote{\code{stpp}} object.
}
  \item{\ldots}{
  }
}
\author{
Robert Clements
}
\seealso{
\code{\link{stpp}}
}
\examples{
data(eq)
eq
}
