\name{print.supresid}
\alias{print.supresid}
\title{
Print details of a \code{supresid} object
}
\description{
Prints the details of the results of finding superposed residuals for a space-time point process.
}
\usage{
print.supresid(x)
}
\arguments{
  \item{x}{
A \code{\dQuote{supresid}} object.
}
}
\details{
Prints the details of the space-time dataset, the superposition rate, the superposed residuals, and the simulated points.
}
\author{
Robert Clements
}
\seealso{
\code{\link{supresid}}
}
\examples{
data(sresiduals1)
sresiduals1
}
